requirements.txt
# Core
openai>=1.0
python-dotenv>=1.0
pydantic>=2.0
pyyaml>=6.0

# Web Scraping
requests>=2.28
beautifulsoup4>=4.11

# Timezone
pytz>=2023.3

# Site Building
markdown>=3.5
